Enum rustc_lexer::LiteralKind

pub enum LiteralKind {
    Int {
        base: Base,
        empty_int: bool,
    },
    Float {
        base: Base,
        empty_exponent: bool,
    },
    Char {
        terminated: bool,
    },
    Byte {
        terminated: bool,
    },
    Str {
        terminated: bool,
    },
    ByteStr {
        terminated: bool,
    },
    CStr {
        terminated: bool,
    },
    RawStr {
        n_hashes: Option<u8>,
    },
    RawByteStr {
        n_hashes: Option<u8>,
    },
    RawCStr {
        n_hashes: Option<u8>,
    },
}

Enum representing the literal types supported by the lexer.

Note that the suffix is not considered when deciding the LiteralKind in this type. This means that float literals like 1f32 are classified by this type as Int. (Compare against rustc_ast::token::LitKind and rustc_ast::ast::LitKind).

Variants

Int

Fields

base: Base

empty_int: bool

“12_u8”, “0o100”, “0b120i99”, “1f32”.

Float

Fields

base: Base

empty_exponent: bool

“12.34f32”, “1e3”, but not “1f32”.

Char

Fields

terminated: bool

“‘a’”, “‘\’”, “‘’‘”, “’;”

Byte

Fields

terminated: bool

“b’a’”, “b’\‘”, “b’‘’”, “b’;”

Str

Fields

terminated: bool

““abc”“, ““abc”

ByteStr

Fields

terminated: bool

“b“abc”“, “b“abc”

CStr

Fields

terminated: bool

c"abc", c"abc

RawStr

Fields

n_hashes: Option<u8>

“r“abc”“, “r#“abc”#“, “r####“ab”###“c”####“, “r#“a”. None indicates an invalid literal.

RawByteStr

Fields

n_hashes: Option<u8>

“br“abc”“, “br#“abc”#“, “br####“ab”###“c”####“, “br#“a”. None indicates an invalid literal.

RawCStr

Fields

n_hashes: Option<u8>

cr"abc", “cr#“abc”#“, cr#"a. None indicates an invalid literal.