Module rustc_lexer::unescape

Utilities for validating string and char literals and turning them into values they represent.

Enums

EscapeError

Errors and warnings that can occur during string unescaping.

Mode

What kind of literal do we parse.

Functions

ascii_check 🔒

byte_from_char 🔒

is_ascii 🔒

scan_escape 🔒

unescape_byte

Takes a contents of a byte literal (without quotes), and returns an unescaped byte or an error.

unescape_byte_literal

Takes a contents of a byte, byte string or raw byte string (without quotes) and produces a sequence of bytes or errors. Values are returned through invoking of the provided callback.

unescape_char

Takes a contents of a char literal (without quotes), and returns an unescaped char or an error

unescape_char_or_byte 🔒

unescape_literal

Takes a contents of a literal (without quotes) and produces a sequence of escaped characters or errors. Values are returned through invoking of the provided callback.

unescape_raw_str_or_byte_str 🔒

Takes a contents of a string literal (without quotes) and produces a sequence of characters or errors. NOTE: Raw strings do not perform any explicit character escaping, here we only translate CRLF to LF and produce errors on bare CR.

unescape_str_or_byte_str 🔒

Takes a contents of a string literal (without quotes) and produces a sequence of escaped characters or errors.