Struct rustc_lexer::cursor::Cursor

pub(crate) struct Cursor<'a> {
    initial_len: usize,
    chars: Chars<'a>,
    prev: char,
}

Peekable iterator over a char sequence.

Next characters can be peeked via first method, and position can be shifted forward via bump method.

Fields
initial_len: usizechars: Chars<'a>

Iterator over chars. Slightly faster than a &str.

prev: char
Implementations

impl<'a> Cursor<'a>

pub(crate) fn new(input: &'a str) -> Cursor<'a>

pub(crate) fn prev(&self) -> char

Returns the last eaten symbol (or '\0' in release builds). (For debug assertions only.)

pub(crate) fn first(&self) -> char

Peeks the next symbol from the input stream without consuming it. If requested position doesn’t exist, EOF_CHAR is returned. However, getting EOF_CHAR doesn’t always mean actual end of file, it should be checked with is_eof method.

pub(crate) fn second(&self) -> char

Peeks the second symbol from the input stream without consuming it.

pub(crate) fn is_eof(&self) -> bool

Checks if there is nothing more to consume.

pub(crate) fn len_consumed(&self) -> u32

Returns amount of already consumed symbols.

pub(crate) fn reset_len_consumed(&mut self)

Resets the number of bytes consumed to 0.

pub(crate) fn bump(&mut self) -> Option<char>

Moves to the next character.

pub(crate) fn eat_while(&mut self, predicate: impl FnMut(char) -> bool)

Eats symbols while predicate returns true or until the end of file is reached.

impl Cursor<'_>

pub(crate) fn advance_token(&mut self) -> Token

Parses a token from the input string.

pub(crate) fn line_comment(&mut self) -> TokenKind

pub(crate) fn block_comment(&mut self) -> TokenKind

pub(crate) fn whitespace(&mut self) -> TokenKind

pub(crate) fn raw_ident(&mut self) -> TokenKind

pub(crate) fn ident_or_unknown_prefix(&mut self) -> TokenKind

pub(crate) fn fake_ident_or_unknown_prefix(&mut self) -> TokenKind

pub(crate) fn number(&mut self, first_digit: char) -> LiteralKind

pub(crate) fn lifetime_or_char(&mut self) -> TokenKind

pub(crate) fn single_quoted_string(&mut self) -> bool

pub(crate) fn double_quoted_string(&mut self) -> bool

Eats double-quoted string and returns true if string is terminated.

pub(crate) fn raw_double_quoted_string(
&mut self,
prefix_len: u32
) -> Result<u8, RawStrError>

Eats the double-quoted string and returns n_hashes and an error if encountered.

pub(crate) fn raw_string_unvalidated(
&mut self,
prefix_len: u32
) -> Result<u32, RawStrError>

pub(crate) fn eat_decimal_digits(&mut self) -> bool

pub(crate) fn eat_hexadecimal_digits(&mut self) -> bool

pub(crate) fn eat_float_exponent(&mut self) -> bool

Eats the float exponent. Returns true if at least one digit was met, and returns false otherwise.

pub(crate) fn eat_literal_suffix(&mut self)

pub(crate) fn eat_identifier(&mut self)