rustc_lexer

Module unescape

Utilities for validating string and char literals and turning them into values they represent.

Enums

EscapeError

Mode

Functions

unescape_byte

unescape_byte_str

Takes a contents of a string literal (without quotes) and produces a sequence of escaped characters or errors.

unescape_char

Takes a contents of a char literal (without quotes), and returns an unescaped char or an error

unescape_raw_byte_str

Takes a contents of a string literal (without quotes) and produces a sequence of characters or errors. NOTE: Raw strings do not perform any explicit character escaping, here we only translate CRLF to LF and produce errors on bare CR.

unescape_raw_str

Takes a contents of a string literal (without quotes) and produces a sequence of characters or errors. NOTE: Raw strings do not perform any explicit character escaping, here we only translate CRLF to LF and produce errors on bare CR.

unescape_str

Takes a contents of a string literal (without quotes) and produces a sequence of escaped characters or errors.