Enum rustc_lexer::unescape::EscapeError

pub enum EscapeError {

    ZeroChars,
    MoreThanOneChar,
    LoneSlash,
    InvalidEscape,
    BareCarriageReturn,
    BareCarriageReturnInRawString,
    EscapeOnlyChar,
    TooShortHexEscape,
    InvalidCharInHexEscape,
    OutOfRangeHexEscape,
    NoBraceInUnicodeEscape,
    InvalidCharInUnicodeEscape,
    EmptyUnicodeEscape,
    UnclosedUnicodeEscape,
    LeadingUnderscoreUnicodeEscape,
    OverlongUnicodeEscape,
    LoneSurrogateUnicodeEscape,
    OutOfRangeUnicodeEscape,
    UnicodeEscapeInByte,
    NonAsciiCharInByte,
    NulInCStr,
    UnskippedWhitespaceWarning,
    MultipleSkippedLinesWarning,

}

Errors and warnings that can occur during string unescaping. They mostly relate to malformed escape sequences, but there are a few that are about other problems.

Variants

ZeroChars

Expected 1 char, but 0 were found.

MoreThanOneChar

Expected 1 char, but more than 1 were found.

LoneSlash

Escaped ’' character without continuation.

InvalidEscape

Invalid escape character (e.g. ‘\z’).

BareCarriageReturn

Raw ‘\r’ encountered.

BareCarriageReturnInRawString

Raw ‘\r’ encountered in raw string.

EscapeOnlyChar

Unescaped character that was expected to be escaped (e.g. raw ‘\t’).

TooShortHexEscape

Numeric character escape is too short (e.g. ‘\x1’).

InvalidCharInHexEscape

Invalid character in numeric escape (e.g. ‘\xz’)

OutOfRangeHexEscape

Character code in numeric escape is non-ascii (e.g. ‘\xFF’).

NoBraceInUnicodeEscape

‘\u’ not followed by ‘{’.

InvalidCharInUnicodeEscape

Non-hexadecimal value in ‘\u{..}’.

EmptyUnicodeEscape

‘\u{}’

UnclosedUnicodeEscape

No closing brace in ‘\u{..}’, e.g. ‘\u{12’.

LeadingUnderscoreUnicodeEscape

‘\u{_12}’

OverlongUnicodeEscape

More than 6 characters in ‘\u{..}’, e.g. ‘\u{10FFFF_FF}’

LoneSurrogateUnicodeEscape

Invalid in-bound unicode character code, e.g. ‘\u{DFFF}’.

OutOfRangeUnicodeEscape

Out of bounds unicode character code, e.g. ‘\u{FFFFFF}’.

UnicodeEscapeInByte

Unicode escape code in byte literal.

NonAsciiCharInByte

Non-ascii character in byte literal, byte string literal, or raw byte string literal.

NulInCStr

UnskippedWhitespaceWarning

After a line ending with ’', the next line contains whitespace characters that are not skipped.

MultipleSkippedLinesWarning

After a line ending with ’', multiple lines are skipped.

Implementations

impl EscapeError

pub fn is_fatal(&self) -> bool

Returns true for actual errors, as opposed to warnings.